requirements.txt
cffi==1.14.0
click==7.1.1
cryptography==2.9
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
jwcrypto==0.7
jwt==1.0.0
MarkupSafe==1.1.1
Pillow==7.1.1
pycparser==2.20
pyrfc==2.0.2
segno==1.0.0
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
